+++ lib/sms_aero/callable.rb
@@ -1,0 +1,6 @@
+module SmsAero::Callable
+  def call(*args)
+    new(*args)
+  end
+  alias [] call
+end