+++ lib/sms_aero/group.rb
@@ -1,0 +1,11 @@
+class SmsAero::Group < String
+  extend SmsAero::Callable
+
+  private
+
+  def initialize(value)
+    group = value.to_s
+    raise "'#{value}' is not a valid group name" unless group[/^\w{1,20}$/]
+    super group
+  end
+end