+++ lib/sms_aero/models/answer.rb
@@ -1,8 +1,0 @@
-class SmsAero
-  class Answer < Evil::Struct
-    attribute :reason, default: proc { nil }
-    attribute :result,
-              Types::FilledString.constructor(&:strip),
-              default: proc { "accepted" }
-  end
-end