+++ lib/sms_aero/models/tariff.rb
@@ -1,8 +1,0 @@
-class SmsAero
-  class Tariff < Evil::Struct
-    attributes type: Types::Coercible::Float, optional: true do
-      attribute :"Direct channel",  as: :direct
-      attribute :"Digital channel", as: :digital
-    end
-  end
-end