+++ lib/sms_aero/response/with_balance.rb
@@ -1,0 +1,5 @@
+class SmsAero
+  class Response::WithBalance < Response
+    option :balance, proc(&:to_f), optional: true
+  end
+end