+++ lib/sms_aero/response/with_groups.rb
@@ -1,0 +1,5 @@
+class SmsAero
+  class Response::WithGroups < Response
+    option :reason, method(:Array), as: :channels, optional: true
+  end
+end