+++ lib/sms_aero/tariff.rb
@@ -1,0 +1,8 @@
+class SmsAero::Tariff
+  extend  Dry::Initializer
+  extend  SmsAero::Callable
+  include SmsAero::Optional
+
+  option :"Direct channel",  proc(&:to_f), optional: true, as: :direct
+  option :"Digital channel", proc(&:to_f), optional: true, as: :digital
+end