+++ lib/sms_aero/types/password.rb
@@ -1,6 +1,0 @@
-module SmsAero::Types
-  Password = \
-    Strict::String.constructor do |value|
-      OpenSSL::Digest::MD5.new.hexdigest(value)
-    end
-end