+++ lib/sms_aero/types/sign_status.rb
@@ -1,9 +1,0 @@
-module SmsAero::Types
-  # Describes statuses of the sign
-  SignStatus = Strict::String.constrained included_in: %w(
-    accepted
-    approved
-    rejected
-    pending
-  )
-end