+++ spec/models/birthday_spec.rb
@@ -1,0 +1,33 @@
+require "spec_helper"
+
+RSpec.describe SmsAero::Birthday do
+  subject { described_class[source] }
+
+  context "date:" do
+    let(:source) { Date.parse "1901-08-17" }
+    it { is_expected.to eq "1901-08-17" }
+  end
+
+  context "datetime:" do
+    let(:source) { DateTime.parse "1901-08-17 10:00:00" }
+    it { is_expected.to eq "1901-08-17" }
+  end
+
+  context "time:" do
+    let(:source) { Time.parse "1901-08-17 10:00:00" }
+    it { is_expected.to eq "1901-08-17" }
+  end
+
+  context "parceable string:" do
+    let(:source) { "1901/08/17" }
+    it { is_expected.to eq "1901-08-17" }
+  end
+
+  context "imparceable string:" do
+    let(:source) { "Foobar" }
+
+    it "fails" do
+      expect { subject }.to raise_error(StandardError, /Foobar/)
+    end
+  end
+end